Add thundra_agent_invocation_sample_onerror Environment Variable This environment variable is added to overcome an issue caused by a deprecated variable, thundra_agent_lambda_sample_timed_out_invocations.
The main functionality is the same as ErrorAwareSampler, however, gives
the ability to provide it without code change, via an environment
variable, to support customers who are using the old variable.

When a sampler and this variable provided at the same time, the sampler
will have the priority and override the environment variable and act
like it's not there.

Within this commit, we also fix an issue in invocation plugin that
were preventing us to receive samples.